+++ previous-versions/accounting/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v0.32.0] previous-versions/accounting
 ### Releases
 - [PyPI v0.32.0] https://pypi.org/project/codat-accounting/0.32.0 - previous-versions/accounting
+
+## 2023-10-30 01:23:52
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.32.1] previous-versions/accounting
+### Releases
+- [PyPI v0.32.1] https://pypi.org/project/codat-accounting/0.32.1 - previous-versions/accounting